AllMemosクラスで、ファイル操作関連部分を別クラスに切り出し
